gym-social-backend/requirements.txt
asgiref==3.8.1
Django==5.0.1
django-cors-headers==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.2
pillow==10.2.0
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
